+++ blueprint/implementations/billing/stripe/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @forklaunch/implementation-billing-stripe
 
+## 1.1.24
+
+### Patch Changes
+
+- chore: update internal packages
+- Updated dependencies
+  - @forklaunch/implementation-billing-base@1.0.24
+  - @forklaunch/interfaces-billing@1.0.23
+
 ## 1.1.23
 
 ### Patch Changes