+++ framework/hyper-express/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @forklaunch/hyper-express
 
+## 1.2.30
+
+### Patch Changes
+
+- Export wrapEmWithTenantContext for tenant based filtering
+- Updated dependencies
+  - @forklaunch/core@1.5.0
+  - @forklaunch/validator@1.2.16
+  - @forklaunch/common@1.2.16
+  - @forklaunch/ws@1.2.30
+
 ## 1.2.29
 
 ### Patch Changes