+++ framework/internal/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @forklaunch/internal
 
+## 1.2.16
+
+### Patch Changes
+
+- Export wrapEmWithTenantContext for tenant based filtering
+- Updated dependencies
+  - @forklaunch/validator@1.2.16
+  - @forklaunch/common@1.2.16
+
 ## 1.2.15
 
 ### Patch Changes